# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "Учебный REST API для заметок (CRUD).",
        "title": "Notes API",
        "contact": {
            "name": "Backend Course",
            "email": "example@university.ru"
        },
        "version": "1.0"
    },
    "basePath": "/",
    "paths": {
        "/notes": {
            "get": {
                "description": "Возвращает список заметок",
                "tags": [
                    "notes"
                ],
                "summary": "Список заметок",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Номер страницы",
                        "name": "page",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "Размер страницы",
                        "name": "limit",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Поиск по title",
                        "name": "q",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/core.Note"
                            }
                        },
                        "headers": {
                            "X-Total-Count": {
                                "type": "integer",
                                "description": "Общее количество"
                            }
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            },
            "post": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "notes"
                ],
                "summary": "Создать заметку",
                "parameters": [
                    {
                        "description": "Данные новой заметки",
                        "name": "input",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/core.NoteCreate"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/core.Note"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
        "/notes/{id}": {
            "get": {
                "tags": [
                    "notes"
                ],
                "summary": "Получить заметку",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/core.Note"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "core.Note": {
            "type": "object",
            "properties": {
                "content": {
                    "type": "string",
                    "example": "Молоко, хлеб, яйца"
                },
                "created_at": {
                    "type": "string",
                    "example": "2023-10-01T12:00:00Z"
                },
                "id": {
                    "type": "integer",
                    "example": 1
                },
                "title": {
                    "type": "string",
                    "example": "Список покупок"
                }
            }
        },
        "core.NoteCreate": {
            "type": "object",
            "properties": {
                "content": {
                    "type": "string",
                    "example": "Текст заметки"
                },
                "title": {
                    "type": "string",
                    "example": "Новая заметка"
                }
            }
        }
    }
}